Modified hadd to run in parallel using a TSelector
Modified hadd to use a TSelector derived class which merges files in parallel
using multiple workers. This can be selected via the -p[numWorkers] flag
when running the program. hadd will invoke a selector with the specified
number of workers. Each worker will merge a fraction of the input files
and the client will merge the result of each worker.
The SelectorArgs class is a container which holds the arguments to be
passed to the workers that will merge the files.

Signed-off-by: Andru Gheorghiu <gheorghiuandru@gmail.com>