rewrite RooRealMPFE to use BidirMMapPipe, take care to make the communications protocol as efficient as possible (as few trips through the pipe as can be achieved)
To this end, the only commands in the parent which flush the buffers to
the other end are the "calculate" command which causes the child to
start the calculation, and the "retrieve result" command which causes
the child to send the result of the calculation. All other commands are
queued in the parent until the "calculate" command is sent, and the
child gets them when all buffers are flushed after the "calculate"
command has been sent as one huge string of requests. In the child, the
only reply which flushes the internal buffers is the reply to "retrieve
result" which sends the result and any evaluation errors that may have
occurred in the child in one go.

The advantage of this mode of operation is that both ends may block in
the OS while trying to read from the pipe until there is a significant
amount of work to be done, resulting in less latency because the OS only
has the schedule the end that becomes runnable once instead of
scheduling it every time the other end sends a tiny request like a
parameter change notification, for example. This makes it explicit that
these "little" requests are nothing more than options to the one
CPU-intensive request which we really want the child process to perform:
the evaluation of the PDF over a certain subset of events.

[FIXME:] One more thought on persistency: This class should be rewritten
somewhat, because persistency is tricky: you cannot persist the child
processes along with the class in the ROOT file and recreate them on the
machine which reads the RooRealMPFE class. That means any reference to
child processes, pids, file descriptors and other communication
structures are meaningless once the class is read back on the target
machine. I can see two ways out: Either we forbid persisting this class
altogether, or we need some special hooks which always write the
persisted version of the class in the Initialize state (irrespective of
the actual state) which will force creation of the child processes on
the target machine when needed. Since the old class had the exact same
issue, I didn't fix this just yet...

Signed-off-by: Wouter Verkerke <verkerke@nikhef.nl>