PROOF: tag identifying additional logfiles
Tags are now used for the Valgrind and IgProf wrappers, generating
additional logfiles. They are appended to the "ordinal" number both for
the master line and the workers lines. Appending tags to ordinals also
incidentally fixes a bug leading to having the additional logfile
concatenated to the ordinary logfile while getting the first.