Replace a constant length vector with a struct
A map of Node* -> vector<double> stores a sum and sort of a squared
sum in the vector, i.e. it has either zero elements (on creation)
or 2 elemnts when used.

There is no need to use a vector here. We can simply use two named
double variables instead. In addition to better runtime performance,
we get better naming and less code for free (remove empty() branch).

Signed-off-by: Danilo Piparo <danilo.piparo@cern.ch>