Allow the use of 1st gen Collection Proxy for bitset and vector<bool>
Because the address of the element of a bitset and vector<bool> can not be
taken, we need to use a temporary storage to implement the interface that
requires the address of the element.  So far this had only been implemented
in the specialized collection proxy (TGenBitsetProxy and TGenVectorBoolProxy).
Those are obtained only when cloning a generic collection proxy.
Consequently using the result TClass::GetCollectionProxy for those two
directly did not work in all cases/interfaces.  However since TTree is making
a copy of the collection proxy, it was correctly in all cases.

This patch resolves all the issues seen in ROOT-8574 including
  - the bitset proxy's iterator did not record it started but still recorded its advance
  - the genertic proxy At function in case of bitset or vector<bool> was not return the right address

We also simplified TGenVectorBoolProxy since we know the actual container type at compilation time.