[TDF] Avoid locking of a weak_ptr in the event loop
This commit removes the need of locking a weak pointer during the event loop,
taking advantage of the fact that we are guaranteed by TDataFrame's logic that
during the event loop TDataFrameImpl must remain in scope.

TDataFrame{Branch,Filter,Action} now store a raw pointer to the TDFImpl object.
TDataFrameInterface stores the weak_ptr<TDFImpl> in their stead.
As it was before, TDataFrameInterface is responsible for checking that the
TDataFrameImpl is still in scope before booking transformations and actions.