[TDF] Replace slot number acquisition mechanism with a lock-free one
the slot number acquisition mechanism was based on a map which had
as key the threadID and as value the slot number. This is not
necessary because every operation is identical irrespective of the
slot - in some sense it is thread-agnostic. A trivial lock-free,
read-only buffer was implemented as helper class of the
TDataFrameImpl::Run method. In addition, the new method allows the
runtime to replace existing threads with new ones: no new entries
in the map are necessary and the total number of slots assigned to
the Operations can stay a constant decided before looping.