Redo setting the title offset default (roottest needs update)
Setting OffsetTitle in the style, leads to the corresponding data member
in the histogram's TAxis to be set.  Since the data member was equal to zero and
is now non-zero this results in a very slight decrease in compression
ratio which is detected by the intentionally strict root/io/filemerger
test (which focus on making sure that the request compression level is
applying consistently).  To accomodate this change, the test needs to
be recalibrated.

This reverts commit a599816620572a880a7b056678f1c349c886ac4a.